Renames <ProjectEdit /> component to <ProjectView /> for more semantic naming; also modifies associated imports into Routes.js
